♻️ 책임분리
