+++ tsconfig.json
@@ -33,8 +33,8 @@
     "src/**/*.vue",
     "tests/**/*.ts",
     "tests/**/*.tsx", 
-    "backend/services/newsService.ts",
-    "backend/services/generatePostService.ts"
+    "backend/services/NewsService.ts",
+    "backend/services/PostService.ts"
   ],
   "exclude": [
     "node_modules"
